[stream-task] Replace std mutex with parking_lot.
