+++ commons/stream-task/Cargo.toml
@@ -16,6 +16,7 @@ futures-retry = "0.5.0"
 pin-project = "0.4.8"
 log = "0.4.14"
 serde = { version = "1.0.123", features = ["derive"] }
+parking_lot = "0.11.1"
 
 [dev-dependencies]
 stest = { path = "../stest" }
